[Decomp][Spell] Cohesion split: extract power/cost to SpellPower.cpp
